moved clearchannel
